+++ packages/datasource-sql/src/connection/index.ts
@@ -1,6 +1,7 @@
 import { Sequelize } from 'sequelize';
 
 import ConnectionOptions from './connection-options';
+import testConnectionWithTimeOut from './connection-tester';
 import handleErrors from './handle-errors';
 import ReverseProxy from './services/reverse-proxy';
 import SequelizeFactory from './services/sequelize-factory';
@@ -47,7 +48,8 @@ export default async function connect(options: ConnectionOptions): Promise<Seque
     }
 
     sequelize = sequelizeFactory.build(await options.buildSequelizeCtorOptions());
-    await sequelize.authenticate(); // Test connection
+
+    await testConnectionWithTimeOut(sequelize, options.debugDatabaseUri);
 
     return sequelize;
   } catch (e) {