+++ packages/forestadmin-client/CHANGELOG.md
@@ -1,3 +1,10 @@
+# @forestadmin/forestadmin-client [1.5.0](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/forestadmin-client@1.4.0...@forestadmin/forestadmin-client@1.5.0) (2023-06-19)
+
+
+### Features
+
+* handle field names containing underscores on update record actions ([#735](https://github.com/ForestAdmin/agent-nodejs/issues/735)) ([d5d7661](https://github.com/ForestAdmin/agent-nodejs/commit/d5d76616e883c51dccf2a5c8e52f8b15d7ff072e))
+
 # @forestadmin/forestadmin-client [1.4.0](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/forestadmin-client@1.3.0...@forestadmin/forestadmin-client@1.4.0) (2023-06-15)